requirements.txt
Flask==3.0.0
APScheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
python-dateutil==2.8.2
Werkzeug==3.0.1
waitress==3.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
